Split off test program out of program.cpp
